design: 게시글 페이지 게시글 컴포넌트 추가 #46
