+++ src/pages/PostPage/style.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  padding: 20px 16px;
+  border-bottom: 1px solid ${({ theme }) => theme.palette.lightGray};
+`;
